refactor: interfaces, code organization
