+++ src/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from './discussion-topic';
+export * from './speed-grader';
