+++ src/interfaces.ts
@@ -1,1 +1,1 @@
-export interface Log extends Pick<Console, 'log' | 'info' | 'warn' | 'error'> {}
+export interface ILog extends Pick<Console, 'log' | 'info' | 'warn' | 'error'> {}
