+++ src/main.ts
@@ -1,12 +1,12 @@
-import { Log } from 'interfaces';
+import { ILog } from './interfaces';
 import { BUILD_ENV } from './constants';
-import { DiscussionTopicHandler, SpeedGraderHandler } from './discussions';
+import { DiscussionTopicHandler, SpeedGraderHandler } from './handlers';
 
 export const VERSION = BUILD_ENV.version;
 export const NAME = BUILD_ENV.name;
 export const PUBLIC_PATH = BUILD_ENV.publicPath;
 
-let log: Log = {
+let log: ILog = {
     log: () => {
         /* empty */
     },
@@ -35,7 +35,7 @@ class AnnotoCanvas {
         log.info('AnnotoCanvas: setup');
 
         const initializeHandlers = (): void => {
-            log.info('AnnotoCanvas: load');
+            log.info('AnnotoCanvas: init handlers');
             discussionTopicHandler.init();
             speedGraderHandler.init();
         };
